requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
dataclasses-json==0.5.7
idna==3.3
isodate==0.6.1
marshmallow==3.15.0
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.3
pyparsing==3.0.7
python-dotenv==0.19.2
python-youtube==0.8.2
requests==2.27.1
requests-oauthlib==1.3.1
schedule==1.1.0
six==1.16.0
spotipy==2.19.0
tabulate==0.8.9
tweepy==4.8.0
typing-inspect==0.7.1
typing_extensions==4.1.1
urllib3==1.26.8
